src/test/java/ma/zyn/app/unit/ws.facade/admin/exigence/ExigenceRestAdminTest.java
package ma.zyn.app.unit.ws.facade.admin.exigence;

import ma.zyn.app.bean.core.exigence.Exigence;
import ma.zyn.app.service.impl.admin.exigence.ExigenceAdminServiceImpl;
import ma.zyn.app.ws.facade.admin.exigence.ExigenceRestAdmin;
import ma.zyn.app.ws.converter.exigence.ExigenceConverter;
import ma.zyn.app.ws.dto.exigence.ExigenceDto;
import org.aspectj.lang.annotation.Before;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class ExigenceRestAdminTest {

    private MockMvc mockMvc;

    @Mock
    private ExigenceAdminServiceImpl service;
    @Mock
    private ExigenceConverter converter;

    @InjectMocks
    private ExigenceRestAdmin controller;

    @Before("")
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }


    @Test
    public void itShouldFindAllExigenceTest() throws Exception {
        // Mock the service to return an empty list
        when(service.findAll()).thenReturn(Collections.emptyList());
        when(converter.toDto(Collections.emptyList())).thenReturn(Collections.emptyList());

        // Call the controller method
        ResponseEntity<List<ExigenceDto>> result = controller.findAll();

        // Verify the result
        assertEquals(HttpStatus.NO_CONTENT, result.getStatusCode());

        // Response body should be empty list
        List<ExigenceDto> responseBody = result.getBody();
        assertNotNull(responseBody);
        assertEquals(0, responseBody.size());
    }

    @Test
    public void itShouldSaveExigenceTest() throws Exception {
        // Mock data
        ExigenceDto requestDto = new ExigenceDto();
        Exigence entity = new Exigence();
        Exigence saved = new Exigence();
        ExigenceDto savedDto = new ExigenceDto();

        // Mock the converter to return the exigence object when converting from DTO
        when(converter.toItem(requestDto)).thenReturn(entity);

        // Mock the service to return the saved client
        when(service.create(entity)).thenReturn(saved);

        // Mock the converter to return the saved exigence DTO
        when(converter.toDto(saved)).thenReturn(savedDto);

        // Call the controller method
        ResponseEntity<ExigenceDto> result = controller.save(requestDto);

        // Verify the result
        assertEquals(HttpStatus.CREATED, result.getStatusCode());

        // Verify the response body
        ExigenceDto responseBody = result.getBody();
        assertNotNull(responseBody);

        // Add assertions to compare the response body with the saved exigence DTO
        assertEquals(savedDto.getReference(), responseBody.getReference());
        assertEquals(savedDto.getLibelle(), responseBody.getLibelle());
        assertEquals(savedDto.getDescription(), responseBody.getDescription());
    }

}
